ServiceTester on same Assembly, to avoid a breaking change
