+++ Gigya.Microdot.Testing.Shared/App.config
