+++ Gigya.Microdot.Testing.Shared/Service/ServiceTesterExtensions.cs
@@ -20,14 +20,25 @@
 // POSSIBILITY OF SUCH DAMAGE.
 #endregion
 
-using System.Reflection;
-using System.Runtime.InteropServices;
+using System;
+using Gigya.Microdot.Hosting.Service;
+using Ninject;
+using Ninject.Parameters;
+using Ninject.Syntax;
 
-// General Information about an assembly is controlled through the following
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("Gigya.Microdot.Testing.Orelans")]
-[assembly: AssemblyProduct("Gigya.Microdot.Testing.Orelans")]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("6d6a62a1-15b5-44c2-ad37-698ab31863e4")]
+namespace Gigya.Microdot.Testing.Shared.Service
+{
+    public static class ServiceTesterExtensions
+    {
+        public static NonOrleansServiceTester<TServiceHost> GetServiceTesterForNonOrleansService<TServiceHost>(this IResolutionRoot kernel, int? basePortOverride = null, TimeSpan? shutdownWaitTime = null)
+            where TServiceHost : ServiceHostBase, new()
+        {
+            NonOrleansServiceTester<TServiceHost> tester = kernel.Get<NonOrleansServiceTester<TServiceHost>>(
+                new ConstructorArgument(nameof(basePortOverride), basePortOverride),
+                new ConstructorArgument(nameof(shutdownWaitTime), shutdownWaitTime)
+            );
+         
+            return tester;
+        }
+    }
+}
