+++ Gigya.Microdot.Testing.Shared/paket.references
