+++ Gigya.Microdot.Testing.Shared/paket.template
