+++ tests/Gigya.Microdot.Orleans.Hosting.UnitTests/AssemblyInitialize.cs
@@ -27,6 +27,7 @@ using Gigya.Microdot.Interfaces.Events;
 using Gigya.Microdot.Orleans.Hosting.UnitTests.Microservice;
 using Gigya.Microdot.ServiceProxy.Caching;
 using Gigya.Microdot.Testing;
+using Gigya.Microdot.Testing.Shared;
 using Ninject.Syntax;
 using NUnit.Framework;
