+++ tests/Gigya.Microdot.Orleans.Hosting.UnitTests/CalculatorServiceTests.cs
@@ -33,7 +33,8 @@ using Gigya.Microdot.Interfaces.HttpService;
 using Gigya.Microdot.Orleans.Hosting.UnitTests.Microservice.CalculatorService;
 using Gigya.Microdot.ServiceProxy;
 using Gigya.Microdot.Testing;
-using Gigya.Microdot.Testing.Orleans.Service;
+using Gigya.Microdot.Testing.Service;
+using Gigya.Microdot.Testing.Shared;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using Ninject;
