+++ tests/Gigya.Microdot.UnitTests/Configuration/TypedConfigTest.cs
@@ -8,6 +8,7 @@ using Gigya.Microdot.Interfaces.Configuration;
 using Gigya.Microdot.SharedLogic.Exceptions;
 using Gigya.Microdot.SharedLogic.Monitor;
 using Gigya.Microdot.Testing;
+using Gigya.Microdot.Testing.Shared;
 using Ninject;
 using NUnit.Framework;
