+++ tests/Gigya.Microdot.UnitTests/Discovery/ServiceDiscoveryConfigChangeTest.cs
@@ -9,7 +9,8 @@ using Gigya.Microdot.ServiceDiscovery;
 using Gigya.Microdot.ServiceDiscovery.Config;
 using Gigya.Microdot.SharedLogic;
 using Gigya.Microdot.Testing;
-using Gigya.Microdot.Testing.Utils;
+using Gigya.Microdot.Testing.Shared;
+using Gigya.Microdot.Testing.Shared.Utils;
 using Ninject;
 
 using NSubstitute;
