+++ tests/Gigya.Microdot.UnitTests/Gigya.Microdot.UnitTests.csproj
@@ -167,9 +167,9 @@
       <Project>{C88DB2A8-A1D2-46F8-8B65-06B9EE3F1662}</Project>
       <Name>Gigya.Microdot.SharedLogic</Name>
     </ProjectReference>
-    <ProjectReference Include="..\..\Gigya.Microdot.Testing\Gigya.Microdot.Testing.csproj">
+    <ProjectReference Include="..\..\Gigya.Microdot.Testing.Shared\Gigya.Microdot.Testing.Shared.csproj">
       <Project>{6D6A62A1-15B5-44C2-AD37-698AB31863E4}</Project>
-      <Name>Gigya.Microdot.Testing</Name>
+      <Name>Gigya.Microdot.Testing.Shared</Name>
     </ProjectReference>
     <ProjectReference Include="..\..\Gigya.ServiceContract\Gigya.ServiceContract.csproj">
       <Project>{DB6D3561-835E-40D5-B9D4-83951CF426DF}</Project>
