+++ tests/Gigya.Microdot.UnitTests/ServiceListenerTests/PortsAllocationTests.cs
@@ -13,6 +13,7 @@ using Gigya.Microdot.ServiceProxy;
 using Gigya.Microdot.SharedLogic;
 using Gigya.Microdot.SharedLogic.Exceptions;
 using Gigya.Microdot.Testing;
+using Gigya.Microdot.Testing.Shared;
 using Gigya.Microdot.UnitTests.ServiceProxyTests;
 
 using Ninject;
