+++ tests/Gigya.Microdot.UnitTests/ServiceProxyTests/AbstractServiceProxyTest.cs
@@ -7,7 +7,7 @@ using Gigya.Microdot.ServiceProxy;
 using Gigya.Microdot.SharedLogic.Events;
 using Gigya.Microdot.SharedLogic.Exceptions;
 using Gigya.Microdot.Testing;
-
+using Gigya.Microdot.Testing.Shared;
 using Metrics;
 
 using Ninject;
